Separate tests and implement test for config.
