+++ test/decorator-test.js
@@ -1,0 +1,136 @@
+/*globals buster:false*/
+buster.testCase("troopjs-compose/decorator", function(run) {
+	"use strict";
+
+	var assert = buster.referee.assert;
+
+	require([
+		"troopjs-compose/mixin/factory",
+		"troopjs-compose/decorator/after",
+		"troopjs-compose/decorator/around",
+		"troopjs-compose/decorator/before",
+		"troopjs-compose/decorator/extend",
+		"troopjs-compose/decorator/from"
+	], function(Factory, after, around, before, extend, from) {
+
+		var Arith = Factory({
+			"sub": function(a, b) {
+				return a - b;
+			},
+			"sub2": function(a, b) {
+				return b - a;
+			}
+		});
+
+		run({
+			"setUp": function() {
+				this.Cal = Factory({
+					name: "cal",
+					supports: {
+						add: true
+					},
+					add: function(a, b) {
+						return a + b;
+					},
+					sub: function(a, b) {
+						return a - b;
+					}
+				});
+			},
+
+			"before": function() {
+				var cal = Factory.create(this.Cal, {
+					// Return array that spreads over arguments of original.
+					add: before(function(a, b) {
+						assert.equals(a, 1);
+						assert.equals(b, 2);
+						return [1, 3];
+					}),
+					// No return value should not interfere original.
+					sub: before(function() { })
+				});
+
+				assert.equals(cal.add(1, 2), 4);
+				assert.equals(cal.sub(2, 1), 1);
+			},
+
+			"after": function() {
+				var cal = Factory.create(this.Cal, {
+					add: after(function(a, b) {
+						assert.equals(1, a);
+						assert.equals(2, b);
+					}),
+					sub: after(function() {
+						return -1;
+					})
+				});
+
+				assert.equals(cal.add(1, 2), 3);
+				assert.equals(cal.sub(2, 1), -1);
+			},
+
+			"around": function() {
+				var cal = Factory.create(this.Cal, {
+					add: around(function(org) {
+						return function(a, b) {
+							a *= a;
+							b *= b;
+							return org(a,b);
+						}
+					})
+				});
+
+				assert.equals(cal.add(1, 2), 5);
+			},
+
+			"from": function() {
+				var cal1 = Factory.create(this.Cal, {
+					sub: from(Arith)
+				});
+
+				var cal2 = Factory.create(this.Cal, {
+					sub: from(Arith, "sub2")
+				});
+
+				var cal3 = Factory.create(this.Cal, {
+					sub: from("add")
+				});
+
+				assert.equals(cal1.sub(2, 1), 1);
+				assert.equals(cal2.sub(2, 1), -1);
+				assert.equals(cal3.sub(2, 1), 3);
+			},
+
+			"extend": function() {
+				var Cal = Factory(this.Cal, {
+					supports: extend({
+						sub: true
+					})
+				});
+
+				var Cal2 = Cal.extend({
+					mul: function(a, b) { return a * b; },
+					supports: extend(function() {
+						return {
+							mul: true,
+							div: false
+						}
+					})
+				});
+
+				var cal1 = Cal();
+				var cal2 = Cal2();
+
+				assert.equals(cal1.supports, {
+					add: true,
+					sub: true
+				});
+
+				assert.equals(cal2.supports, {
+					div: false,
+					mul: true
+				});
+			}
+		});
+	});
+});
